Add factory to GroupViewerCombinator, remove from GroupLogic
