+++ app/user/model/GroupViewerCombinator.scala
@@ -27,3 +27,24 @@ package user.model
  * @param maintainers groups that can delete, administrate of migrate the target
  */
 case class GroupViewerCombinator(viewers: Set[Group], editors: Set[Group], maintainers: Set[Group])
+
+object GroupViewerCombinator {
+
+  def fromRelations(relations: Seq[(Group, Viewer)]): GroupViewerCombinator = {
+    var maintainers: Set[Group] = Set()
+    var editors: Set[Group] = Set()
+    var viewers: Set[Group] = Set()
+    relations.foreach(rel => {
+      val (group, viewer) = rel
+      if(ViewerRole.isAtLeastMaintainer(viewer.role)){
+        maintainers = maintainers + group
+      } else if(ViewerRole.isAtLeastEditor(viewer.role)){
+        editors = editors + group
+      } else if(ViewerRole.isAtLeastViewer(viewer.role)){
+        viewers = viewers + group
+      }
+    })
+    GroupViewerCombinator(viewers, editors, maintainers)
+  }
+
+}
