+++ app/user/service/GroupLogic.scala
@@ -18,25 +18,6 @@
 
 package user.service
 
-import user.model.{Group, GroupViewerCombinator, Viewer, ViewerRole}
-
 object GroupLogic extends ViewerProcessor {
-
-  def buildGroupViewerCombinator(relations: Seq[(Group, Viewer)]): GroupViewerCombinator = {
-    var maintainers: Set[Group] = Set()
-    var editors: Set[Group] = Set()
-    var viewers: Set[Group] = Set()
-    relations.foreach(rel => {
-      val (group, viewer) = rel
-      if(ViewerRole.isAtLeastMaintainer(viewer.role)){
-        maintainers = maintainers + group
-      } else if(ViewerRole.isAtLeastEditor(viewer.role)){
-        editors = editors + group
-      } else if(ViewerRole.isAtLeastViewer(viewer.role)){
-        viewers = viewers + group
-      }
-    })
-    GroupViewerCombinator(viewers, editors, maintainers)
-  }
 
 }
